Switch to hyprpanel from waybar
Essentially, this is done so that I can make quicker progress on my
desktop. The waybar configuration was taking just a little too long for
my comfort. Eventually, I think I'm going to want to take the time to
work on it, but for now, this will get me 90% of the way there.