Use app2unit and improve swww startup
app2unit seems to make things much faster to launch, so it seems like
it's a good idea to use by default in most places that we want to launch
some app that is long-running.

Note that when using app2unit, you must have the env variables set. It
does seem to work when you don't have them set, but ... weird things
happen when it's unset. For instance, the wm hangs when you "log out".
It also sometimes seemingly has issues shutting down, I think. So yeah,
the env edits are necessary.